Strip out a huge amount of unused code from miriad.py.
I haven't used any of the mutation / predecessor stuff and
I haven't touched the vis.fAnt () stuff in a long time either.
Just dump the code so I don't have to document it. The interactive
helpers could be added to new XVisData and XImData classes if
I ever want to resurrect them.