mirexec,miriad: massively clean up task-launching infrastructure
There was a lot of rickety old code for the launching of MIRIAD
tasks, and much of it came from back when I was experimenting
with doing most data analysis in an IPython shell.

Other parts of the code just weren't well-designed. Task
objects, which really store information about launching tasks,
were overloaded to also store information about currently-running
subprocesses. The API was not amenable to launching tasks and
not blocking on them, losing potential parallelism wins for many
operations.

Clean up a bunch of that stuff and improve it. Provide a subclass
of subprocess.Popen that provides a slightly saner API for
standard usage by clients. Just delete lots of code that hasn't
been used for years.