mirtask.emucal: new module emulating MIRIAD's calibration logic
MIRIAD's built-in calibration routines should be used whenever
possible, but sometimes it's necessary to emulate their behavior
to achieve certain effects in Python (e.g., reverse-apply
calibration changes, or emulate them for data that aren't in
the MIRIAD format). Start providing that functionality here.

The goal is complete compatibility with MIRIAD. A bit of testing
suggests that things work well with the gains code. We should
really set up a real testing framework to test these things
in detail, since compatibility is very, very important here.