mirtask: rework user-friendly exception handling
In the course of work on ARF I iterated the user-friendly
exception handler a bit more. The SystemExit exception pretty
much behaves by default in the way I wanted ProgramFailError
to, so use that rather than defining my own exception. Add
a die() function in mirtask.util that provides the formatting
sugar.