mirtask/util.py: tighten up basepol code a whole lot
Rename "basepol" to "32-bit packed basepol" and be more aggressive
about this in the function naming. Most of my code that works with
basepols uses 2-tuples of antpols, and I want to be able to use
the word "basepol" to refer to those things. Also improve the
argument and consistency checking to be more correct and sensible.