mirtask: 64-bit correctness with int32 vs int64
I was a bit confused because in C (i.e., in MIRIAD), an int is
32 bits on a 64-bit machine, while in Numpy, N.int is a 64 bits.
It turns out that this is due to some Python conventions, and
Numpy helpfully provides an "N.intc" type for getting C-sized
integers in the Python domain. This lets us deal with a few bits
of MIRIAD code that require what I call "plain ints" in a
reasonable way in the Python world. Woo.