mirtask/__init__.py: use xyopen mode "append" for "rw" rather than "old"
It turns out that when you use xyopen to open an image in "old" mode,
the image data are actually read-only. This is a pleasant change from
most permissions-checking in MIRIAD, but we were calling this "rw"
mode.

xyopen "append" mode seems to be a better match to "rw". As far as I
can tell, this is equivalent to "old" mode, except that the "image"
item is opened in read-write mode and the file position is initially
set to the end of the item. This latter bit is irrelevant, I think,
because we always need to seek first before doing any image I/O
operations.

We could add an "r" mode that mapped to xyopen "old" in the future.
I'm reticent about this though since MIRIAD lets you do a lot of write
operations on datasets opened in "old" mode.