mirtask/__init__.py: enable WCS loading of images created by the 'fft' task
The 'fft' task writes images with main axes called UU---SIN (e.g.) and
VV---SIN. wcslib looks at the -SIN suffix and tries to treat the axes
as celestial ones, but errors out because it (rightfully) can't
interpret UU or VV as longitudes or latitudes. To work around this, we
modify the ctype presented to wcslib to make it not look like a
celestial axis ctype -- specifically, we chop off the projection
(-SIN) part.