mirtask/__init__.py: prevent segfault when opening many images
I was getting semi-mysterious segfaults in programs that had many
images open simultaneously -- though the segfaults were happening on
image close, which was usually after the programs had done their work,
so they weren't showstoppers.

It turns out that this is because we were requesting too many image
axes on image open. For each image MIRIAD has a statically-allocated
buffer of axis lengths, and xyopen_c doesn't check the naxis argument
to see if it's larger than the static buffer size, MAXNAX. This causes
the image opening code to scribble off the end of the axis buffer and
overwrite other memory. It's not clear to me why this was only a
problem when opening many images, but it's definitely the problem.

I'll fix MIRIAD to do the right thing when naxis > MAXNAX. In the
meantime, request 7 axes (the default value of MAXNAX), rather than
16. This solves the crash (unless someone uses a different smaller
value of MAXNAX of course .... graaaaaaaaahhhhhhh static buffers!!!)