mirtask/_miriad_cmodule.c: update xyz{s2c,c2s} prototypes
The signatures for these functions changed, with the subcubenr
parameter going from "int" to "long int" type.

Note that this means that the MIRIAD subroutine library is breaking
binary compatibility, which is lame. This is certainly not the first
instance in which this has happened and realistically I don't think it
will be possible to prevent other developers from continuing to do
this.

In theory, we should look at the prototype of broken-compat functions
like these and figure out whether they're (e.g.) int or long-int at
configure time. This is an enormous pain in the ass and I don't want
to do it. The current policy is to track CVS HEAD of MIRIAD as best we
can. (Looking at the prototypes would also be dangerous if the header
files and C source files got out of sync, which honestly sounds like a
perfectly likely scenario.)