doc/conf.py: hack to see if I can get this working on readthedocs.org
I think readthedocs is having problems because we don't have conf.py. I'm
pretty sure this won't work because I think it's going to want to build
the full package, and that's not going to be possible because we don't have
MIRIAD packaged up on PyPI or anything ... but it's worth checking if this
is enough to get us partway.